pyproject.toml
[project]
name = "riksbanken-rates"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
    "ruff>=0.12.3",
]
